+++ CHANGELOG.md
@@ -4,16 +4,25 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Improve telemetry to uniformly log all errors for reconnectable chat and reconnect availability
+
 ## [0.5.6]
+
 ### Fixed
+
 - Fix error propagation when a call to an endpoint timesout
 - Fix unit tests
 
 ### Added
+
 - Error propagation when getchattoken returns an empty dashboard
 
 ## [0.5.5]
+
 ### Security
+
 - Uptake [axios@1.7.4](https://www.npmjs.com/package/axios/v/1.7.4)
 
 ## [0.5.3] - 2024-06-18